requirements.txt
Django
djangorestframework
celery
redis
PyPDF2